Implement StatsBar and DominanceChartSection components in RoundsAPage
- Added StatsBar component to display key statistics with responsive styling based on compact mode.
- Introduced DominanceChartSection for visual representation of dominance data.
- Updated RoundsAPage layout to incorporate new components and handle scroll events for compact mode.
- Adjusted InfoBar to conditionally render StatsBar and DominanceChartSection based on the compact state.
